refact: rename showcase package
