+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/showcase/components/BadgeShowcase.kt
@@ -1,4 +1,4 @@
-package com.github.felipecastilhos.pokedexandroid.uicatalog.components
+package com.github.felipecastilhos.pokedexandroid.showcase.components
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
