+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/showcase/components/DividerShowcase.kt
@@ -1,4 +1,4 @@
-package com.github.felipecastilhos.pokedexandroid.uicatalog.components
+package com.github.felipecastilhos.pokedexandroid.showcase.components
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
