+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/showcase/theme/ColorShowcase.kt
@@ -1,4 +1,4 @@
-package com.github.felipecastilhos.pokedexandroid.uicatalog
+package com.github.felipecastilhos.pokedexandroid.showcase
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
